Update join predicate tests to use int64 for integer fields
This commit modifies the createTestTuple function in join_predicate_test.go and join_test_mock.go to replace int32 with int64 for integer fields. This change ensures consistency with recent updates across the codebase for improved integer handling.

+++ pkg/execution/join/join_predicate_test.go
@@ -133,7 +133,7 @@ func createTestTuple(fieldTypes []types.Type, values []any) *tuple.Tuple {
 	for i, val := range values {
 		var field types.Field
 		switch v := val.(type) {
-		case int32:
+		case int64:
 			field = types.NewIntField(v)
 		case string:
 			field = types.NewStringField(v, types.StringMaxSize)
