+++ pkg/execution/join/join_test_mock.go
@@ -94,7 +94,7 @@ func createJoinTestTuple(tupleDesc *tuple.TupleDescription, values []interface{}
 	for i, val := range values {
 		var field types.Field
 		switch v := val.(type) {
-		case int32:
+		case int64:
 			field = types.NewIntField(v)
 		case string:
 			field = types.NewStringField(v, types.StringMaxSize)
